Extract focus card module
