+++ tests/test-a11y-phase3.js
@@ -32,6 +32,7 @@ console.log('=== Phase 3 A11y Tests ===\n');
 
   // ─── 2. Clickable divs gain role+tabindex ───
   const viewsSrc = read('/js/views.js');
+  const focusCardSrc = read('/js/focus-card.js');
   const markerDetailSrc = read('/js/marker-detail-modal.js');
   const dashboardRenderersSrc = read('/js/dashboard-widget-renderers.js');
   assert('chart-card has role and tabindex',
@@ -51,7 +52,7 @@ console.log('=== Phase 3 A11y Tests ===\n');
   assert('ref-editable span has role+tabindex',
     markerDetailSrc.includes('class="ref-editable" role="button" tabindex="0"'));
   assert('focus-card refresh has aria-label',
-    viewsSrc.includes('class="focus-card-refresh" onclick="refreshFocusCard()" aria-label="Regenerate insight"'));
+    focusCardSrc.includes('class="focus-card-refresh" onclick="refreshFocusCard()" aria-label="Regenerate insight"'));
 
   const cycleSrc = read('/js/cycle.js');
   assert('cycle-prompt is a semantic button',
