+++ tests/test-correctness-phase2.js
@@ -80,6 +80,7 @@ const pwaAppShellAssets = [
   '/js/dashboard-widgets.js',
   '/js/dashboard-widget-controls.js',
   '/js/dashboard-widget-renderers.js',
+  '/js/focus-card.js',
   '/js/light-conditions-now.js',
   '/js/light-sessions-view.js',
   '/js/compare-correlations.js',
